signin-signup-service/requirements.txt
# libraries
mysql-connector==2.2.9
mysql-connector-python==9.2.0
flask==3.1.0
bcrypt-flask==1.0.2
flask-jwt-extended==4.7.1
flask-cors==5.0.0
